% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Bayenet.R
\name{print.Bayenet.pred}
\alias{print.Bayenet.pred}
\title{print a predict.Bayenet object}
\usage{
\method{print}{Bayenet.pred}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{predict.Bayenet object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Print a summary of a predict.Bayenet object
}
\seealso{
\code{\link{predict.Bayenet}}
}
