% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Bayenet.R
\name{print.Bayenet}
\alias{print.Bayenet}
\title{print a Bayenet object}
\usage{
\method{print}{Bayenet}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{Bayenet object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Print a summary of a Bayenet object
}
\seealso{
\code{\link{Bayenet}}
}
