% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Bayenet.R
\name{print.Selection}
\alias{print.Selection}
\title{print a Selection object}
\usage{
\method{print}{Selection}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{Selection object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\value{
No return value, called for side effects.
}
\description{
Print a summary of a Selection object
}
\seealso{
\code{\link{Selection}}
}
